+++ src/common/commonTTS.ts
@@ -30,5 +30,5 @@ export abstract class CommonTTS<T extends CommonTTSRequest> {
         this.player = player;
     }
 
-    abstract getRequest(text: string, voice?: string): T;
+    abstract getRequest(text: string, voice?: string | null): T;
 }